tools/recon/pyproject.toml
[project]
name = "recon"
version = "0.7.0"
description = "Mechanical collection system — collect, normalize, query via parquet + DuckDB"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Yashodeep Vyas", email = "yza.vyas@gmail.com" }
]

dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "xmltodict>=0.13",
    "duckdb>=1.0",
    "glom>=23.0",
    "pymupdf>=1.24",
    "markdownify>=0.14",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
recon = "recon.adapters._in.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recon"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/recon/**/*.py",
    "src/recon/configs/*.yaml",
]
